neurodecode_a2a/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
google-adk==0.4.0
google-genai>=1.0.0
google-cloud-firestore==2.14.0
python-dotenv==1.0.1
httpx>=0.28.1,<1.0.0
